fix: set LOG_BUFFER_MAX to 4
Before, I assumed we wouldn't include the buffer size information in an
error report. But considering that someone may want to enter a '0' as
debug information, we should include it even though it takes up valuable
space.

So now you can only store up to 4 bytes of debug data. Space is very
limited!